requirements.txt
absl-py==1.3.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bidict==0.22.0
blinker==1.4
cachetools==5.2.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
contourpy==1.0.7
cryptography==3.4.8
cycler==0.11.0
defusedxml==0.7.1
detecto==1.2.2
distro==1.7.0
dnspython==2.1.0
fasteners==0.14.1
flatbuffers==22.12.6
fonttools==4.38.0
frozenlist==1.3.1
future==0.18.2
gast==0.4.0
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-aiplatform==1.19.1
google-cloud-bigquery==2.34.4
google-cloud-core==2.3.2
google-cloud-resource-manager==1.6.3
google-cloud-storage==2.7.0
google-cloud-vision==3.1.4
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpc-google-iam-v1==0.12.4
grpcio==1.51.1
grpcio-status==1.48.2
h5py==3.7.0
httplib2==0.20.2
idna==2.10
importlib-metadata==4.6.4
importlib-resources==5.10.2
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jstyleson==0.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.4.4
labelImg==1.8.6
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.6
libpyfoscam==1.2.2
lockfile==0.12.2
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.0
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.2
munkres==1.1.4
natsort==8.2.0
netifaces==0.10.6
networkx==2.8.8
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
olefile==0.46
opencv-python==4.6.0.66
openvino==2022.3.0
openvino-dev==2022.3.0
openvino-telemetry==2022.3.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.3.0
proto-plus==1.22.1
protobuf==3.19.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyRFC3339==1.1
python-dateutil==2.8.1
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.0
SecretStorage==3.3.3
six==1.16.0
soupsieve==2.3.2.post1
texttable==1.6.7
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
wadllib==1.3.6
yarl==1.8.2
zipp==3.13.0
